Split views module into a handful of submodules.
We'll be adding more logic (related to agency searches), so it seemed like a
good time to split the views.py module up.
